bdb_fsm_app/requirements.txt
asn1crypto==0.24.0
base58==1.0.3
bigchaindb-driver==0.6.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==6.7
contourpy==1.1.1
cryptoconditions==0.8.0
cryptography==2.3.1
cycler==0.12.1
Flask==1.0.2
Flask-Cors==3.0.6
fonttools==4.44.0
graphviz==0.20.1
idna==2.7
importlib-resources==6.1.1
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.4.5
MarkupSafe==1.0
matplotlib==3.7.3
numpy==1.24.4
packaging==23.2
Pillow==10.1.0
pyasn1==0.4.8
pycparser==2.18
pydot==1.4.2
pygraphviz==1.11
pymongo==3.7.0
PyNaCl==1.1.2
pyparsing==3.1.1
pysha3==1.0.2
python-dateutil==2.8.2
python-rapidjson==0.6.3
python-rapidjson-schema==0.1.1
requests==2.27.1
six==1.11.0
transitions==0.9.0
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
zipp==3.17.0
